requirements.txt
django>=4.2
djangorestframework
djangorestframework-simplejwt
django-storages
boto3
